benchmark/02-langchain-rag-agent/requirements.txt
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-community>=0.3.0,<0.4.0
faiss-cpu>=1.8.0
requests>=2.31.0
python-dotenv>=1.0.0
